requirements.txt
### 2. requirements.txt
```
flask
psycopg2-binary
```
